"Add Panopto Video" button plugin for Moodle's TinyMCE editor
Adds the ability to embed selected panopto videos in a TinyMCE editor
instance within moodle. Clicking the button in the editor brings up a
dialog allowing a Panopto folder to be selected (or with one already
selected if the id is specified), from which one or more videos can be
marked for insertion into the editor.